% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_dataframes.R
\name{add_num_result}
\alias{add_num_result}
\title{Create a Numeric Results Variable}
\usage{
add_num_result(df)
}
\arguments{
\item{df}{The dataframe to add the \code{Conc} variable to. Dataframe
must have variables named \code{Result}, \code{MDL}, and \code{RL}.
The character \code{Result} variable must have non-detect values
identified as either "< RL" or "< MDL". \code{MDL} and \code{RL}
must be "numeric" class.}
}
\value{
A new variable named \code{Conc}, which is a clone of the
    \code{Result} variable converted to "numeric" class.
}
\description{
Creates a new variable named \code{Conc}, which is a clone
    of the \code{Result} variable converted to "numeric" class. This
    function substitutes the Method Detection Limit (MDL) or Reporting
    Limit (RL) for the non-detect values contained in the \code{Result}
    variable.
}
