% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_dataframes.R
\name{add_short_sta_names}
\alias{add_short_sta_names}
\title{Create a Variable of Short Station Names}
\usage{
add_short_sta_names(df)
}
\arguments{
\item{df}{The dataframe to add the \code{ShortName} variable to. Dataframe
    must have a character variable named \code{StationName}. Each
    observation under the \code{StationName} variable must have one of the
    following values:
\itemize{
\item Cache Slough near Ryer Island
\item CCSB- Low Flow Channel
\item CCSB Overflow Weir- North
\item CCSB Overflow Weir- South
\item Fremont Weir- East Side
\item Fremont Weir- Middle
\item Fremont Weir- West Side
\item Knights Landing Ridge Cut
\item Liberty Cut below Stairsteps
\item Miner Slough near Sac River
\item Prospect Slough
\item Putah Creek at Mace Blvd
\item Sac River above the Sacramento Weir
\item Shag Slough below Stairsteps
\item Toe Drain at 1/2 Lisbon
\item Toe Drain at County Road 22
\item Toe Drain at Interstate 80
\item Toe Drain at Lisbon Weir
}}
}
\value{
A new character variable named \code{ShortName}, which contains
    shortened station names of the \code{StationName} variable.
}
\description{
Creates a new variable named \code{ShortName}, which contains
    shortened station names of the \code{StationName} variable. These
    short station names will be used in plots of the data.
}
