% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{color_pal_owhg}
\alias{color_pal_owhg}
\title{Custom Color Palette for Report Figures}
\usage{
color_pal_owhg(num, aes_type = c("fill", "color"))
}
\arguments{
\item{num}{An integer representing the number of colors for the
palette.}

\item{aes_type}{The aesthetic type to apply the palette to. Can be
either \code{"fill"} or \code{"color"}.}
}
\value{
A ggplot object with a custom color palette applied to it
    with the specified number of colors and aesthetic type.
}
\description{
Applies a custom color palette for figures in the final
    report for the Open Water Mercury studies. The palette is based
    on the Viridis palette. Function is to be used on ggplot objects.
}
