% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{conv_fact_inlet_names}
\alias{conv_fact_inlet_names}
\title{Apply Factor Order to Inlet Stations}
\usage{
conv_fact_inlet_names(df)
}
\arguments{
\item{df}{The dataframe to apply the function to. Dataframe must have
    a character variable named \code{StationName}. Each observation under
    the \code{StationName} variable must have one of the following
    values:
\itemize{
\item CCSB
\item Fremont Weir
\item KLRC
\item Putah Creek
\item Sacramento Weir
}}
}
\value{
A dataframe with the \code{StationName} variable converted to
    a factor
}
\description{
Converts the \code{StationName} variable from a character
    variable to a factor variable to apply a consistent order in
    plots of the inlet station data. The \code{StationName} variable
    contains a standardized set of inlet station names only.
}
