% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{conv_fact_samplingevent}
\alias{conv_fact_samplingevent}
\title{Convert \code{SamplingEvent} Variable to a Factor}
\usage{
conv_fact_samplingevent(df)
}
\arguments{
\item{df}{The dataframe to apply the function to. Dataframe must have
    a character variable named \code{SamplingEvent}. Each observation
    under the \code{SamplingEvent} variable must have one of the
    following values:
\itemize{
\item Dec 22-23, 2014
\item Mar 15-16, 2016
\item Jan 11-12, 2017
\item Jan 24-25, 2017
\item Jan 31-Feb 1, 2017
\item Feb 14-15, 2017
\item Mar 1-2, 2017
\item Mar 15-16, 2017
\item Mar 28-29, 2017
\item Apr 11-12, 2017
\item Apr 25-26, 2017
}}
}
\value{
A dataframe with the \code{SamplingEvent} variable converted
    to a factor
}
\description{
Converts the \code{SamplingEvent} variable from a
    character variable to a factor variable to apply a consistent
    order in plots of the data. The \code{SamplingEvent} variable
    contains a standardized set of sampling event names.
}
