% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{conv_fact_short_sta_names}
\alias{conv_fact_short_sta_names}
\title{Convert \code{ShortName} Variable to a Factor}
\usage{
conv_fact_short_sta_names(df)
}
\arguments{
\item{df}{The dataframe to apply the function to. Dataframe must have
    a character variable named \code{ShortName}. Each observation under
    the \code{ShortName} variable must have one of the following
    values:
\itemize{
\item Cache Sl
\item CCSB- LFC
\item CCSB- OW N
\item CCSB- OW S
\item Fremont Wr- E
\item Fremont Wr- M
\item Fremont Wr- W
\item KLRC
\item Liberty Cut
\item Miner Sl
\item Prospect Sl
\item Putah Ck
\item Shag Sl
\item SR above the Sac Wr
\item Toe Dr at 1/2 Lisbon
\item Toe Dr at I-80
\item Toe Dr at Lisbon
\item Toe Dr at Rd 22
}}
}
\value{
A dataframe with the \code{ShortName} variable converted to
    a factor
}
\description{
Converts the \code{ShortName} variable from a character
    variable to a factor variable to apply a consistent order in
    plots of the data. The \code{ShortName} variable contains a
    standardized set of shorter station names.
}
