% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field_data}
\alias{field_data}
\title{Field Measurements}
\format{data frame with 165 rows and 7 columns
\describe{
\item{StationName}{The station name where the measurement was collected}
\item{SampleDate}{The sampling date of the field measurement}
\item{CollectionTimePST}{The collection time of the field measurement in PST}
\item{WaterTempC}{The water temperature in degrees Celcius}
\item{SpCond}{The specific conductance in uS/cm}
\item{DissOxy}{The dissolved oxygen concentration in mg/L}
\item{Turbidity}{The turbidity in NTU}
}}
\usage{
field_data
}
\description{
Contains field measurements collected for the Yolo Bypass
    Mass Balance Study. Measurements include Water Temperature, Specific
    Conductance, Dissolved Oxygen, and Turbidity.
}
\keyword{datasets}
