% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_dataframes.R
\name{mod_result}
\alias{mod_result}
\title{Create a New Variable \code{Conc}}
\usage{
mod_result(df)
}
\arguments{
\item{df}{The dataframe to add the \code{Conc} variable to. Dataframe
must have variables named \code{Result}, \code{MDL}, and \code{RL}.}
}
\value{
A new variable named \code{Conc}, which is a clone of the
    \code{Result} variable converted to "numeric" type.
}
\description{
Creates a new variable named \code{Conc}, which is a clone
    of the \code{Result} variable converted to "numeric" type. This
    function substitutes the Method Detection Limit (MDL) or Reporting
    Limit (RL) for the non-detect values contained in the \code{Result}
    variable.
}
