% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwaterhg-package.R
\docType{package}
\name{openwaterhg-package}
\alias{openwaterhg}
\alias{openwaterhg-package}
\title{openwaterhg: Analysis Tools For DWR's Open Water Mercury Studies}
\description{
A collection of commonly-used R functions and datasets that are used for data analysis and visualizations of DWR's Open Water Mercury Studies. These studies include the Yolo Bypass Mass Balance Study and three Vegetation Senescence studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mountaindboz.github.io/openwaterhg}
  \item \url{https://github.com/mountaindboz/openwaterhg}
  \item Report bugs at \url{https://github.com/mountaindboz/openwaterhg/issues}
}

}
\author{
\strong{Maintainer}: Dave Bosworth \email{David.Bosworth@water.ca.gov} (\href{https://orcid.org/0000-0003-0740-3390}{ORCID})

}
\keyword{internal}
