% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{part_conc_calc}
\alias{part_conc_calc}
\title{Calculated Particulate Concentrations}
\format{
data frame with 488 rows and 6 columns
\describe{
\item{StationName}{The station name}
\item{SampleDate}{The sampling date}
\item{CollectionTime}{The sample collection time in PST}
\item{Analyte}{The name of the calculated particulate parameter}
\item{Conc}{The calculated particulate concentration}
\item{Units}{The units of the \code{Conc} variable}
}
}
\usage{
part_conc_calc
}
\description{
Contains calculated particulate concentration data for mercury,
    methylmercury, and organic carbon for the Yolo Bypass Mass Balance Study.
    Particulate concentrations were calculated as the difference between the
    unfiltered and filtered concentrations. The calculations for this dataset
    are in the "YB_Mass_Balance/Concentrations/Calculate_Particulate_Fractions.R"
    script file.
}
\keyword{datasets}
