% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{qual_codes_flags}
\alias{qual_codes_flags}
\title{Qual Codes and Flags}
\description{
Following is more information on the Qual Codes and Flags used in the
concentration datasets in the \code{openwaterhg} package.
\subsection{Qual Codes}{

\describe{
\item{\strong{J}}{Not valid, but the flaw was determined to not be detrimental by best
professional judgment; result can be used but with caution}
\item{\strong{R}}{Not valid, and existing information indicates that the result should
\strong{not} be used in any analysis}
}
}

\subsection{Flags}{

\describe{
\item{BD}{Detection in Blank Sample: concentration in the blank sample was greater
than 20\% of the concentration of the associated ambient or environmental sample(s).}
\item{DNQ}{Detected Not Quantifiable: value is greater than the Method Detection Limit
(MDL) but less than the Reporting Limit (RL). Considered to be an estimate.}
\item{FV}{Field Variability: the Relative percent difference (RPD) for associated field
duplicate was greater than the Measurement Quality Objective of 25\% (30\% for VSS,
TOC and DOC). Only applies when one or more of the sample results are greater than 10
times the RL or MDL (for parameters with MDL's provided).}
\item{FGT}{Value of the filtered sample was greater than its associated total sample.}
\item{NRS}{Sample may not have been representative of waterbody. Reason explained in
\code{MME_Comments}.}
}
}
}
\keyword{internal}
