% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summ_stat}
\alias{summ_stat}
\title{Calculate Descriptive Summary Statistics}
\usage{
summ_stat(df, data_var, group_var1, group_var2 = NULL)
}
\arguments{
\item{df}{The dataframe with the data to calculate summary statistics on}

\item{data_var}{The variable in \code{df} that contains the numerical
data to summarize. Must be "numeric" class.}

\item{group_var1}{The mandatory first (or only) grouping variable in
\code{df}.}

\item{group_var2}{The optional second grouping variable in \code{df}.}
}
\value{
A dataframe with the following summary statistics for either
    one (\code{group_var1}) or two (\code{group_var1} and
    \code{group_var2}) grouping variables:
\itemize{
\item "N" - Sample Size
\item "Mean" - Average Value
\item "StDev" - Standard Deviation
\item "Minimum" - Minimum Value
\item "Q1" - 25 percent quantile
\item "Median" - Median Value
\item "Q3" - 75 percent quantile
\item "Maximum" - Maximum Value
\item "IQR" - Interquartile Range (Q3 - Q1)
}
}
\description{
Calculates descriptive summary statistics for one or two
    grouping variables.
}
