% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conc_data}
\alias{conc_data}
\title{Concentration Data in Water}
\format{
data frame with 3,398 rows and 12 columns
\describe{
\item{SampleCode}{The unique identifier for each sample produced by Bryte Lab}
\item{StationName}{The station name where the water sample was collected}
\item{SampleDate}{The sampling date}
\item{CollectionTime}{The sample collection time in PST}
\item{Analyte}{The analytical parameter measured}
\item{Result}{The analytical result. "< RL" or "< MDL" indicates that the result
was less than the Reporting Limit or Method Detection Limit, respectively.
The results for samples that have associated field duplicates or lab replicates
are the average of the two values.}
\item{RL}{The analytical reporting limit of the sample}
\item{MDL}{The analytical method detection limit of the sample}
\item{Units}{The units of the result}
\item{LabComments}{Comments provided by the analytical laboratory}
\item{MME_Comments}{Comments provided by staff from the DWR Mercury Monitoring and
Evaluation section}
\item{QualCode}{Qualification code(s) for the sample. See
\link[=qual_codes_flags]{Qual Codes and Flags} for more information.}
}
}
\usage{
conc_data
}
\description{
Contains all water concentration data collected for the Yolo Bypass
Mass Balance Study which were used for load estimates and other analyses. This
dataset only contains data for normal samples; field QA data (blanks and duplicates)
and laboratory QA data (replicates, laboratory control samples, blanks, matrix spikes)
are excluded from this dataset. Field QA data are provided in \code{\link{qa_field_blanks}} and
\code{\link{qa_field_dups}}. The raw data for this dataset were integrated, cleaned, and QA'ed
in the \href{https://github.com/mountaindboz/Open-Water-Hg-Studies/blob/master/YB_Mass_Balance/Concentrations/Compile_and_Clean_Conc_Data_FloodEvents.R}{YB_Mass_Balance/Concentrations/Compile_and_Clean_Conc_Data_FloodEvents.R}
script file.
}
\keyword{datasets}
