% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conc_data}
\alias{conc_data}
\title{Concentration Data in Water}
\format{data frame with 3,582 rows and 13 columns
\describe{
\item{SampleCode}{The unique identifier for each sample produced by Bryte Lab}
\item{StationName}{The station name where the water samples were collected}
\item{SampleDate}{The sampling date}
\item{CollectionTimePST}{The sample collection time in PST}
\item{Analyte}{The analytical parameter measured}
\item{Result}{The analytical result}
\item{ResQual}{A numeric variable that represents whether the result was below
    the detection limit (RL or MDL). A value of "0" indicates that the result
    was detected, a value of "1" indicates that the result was below the
    detection limit, and a value of "0.5" indicates that either the replicate
    or parent sample was below the detection limit.}
\item{RL}{The analytical reporting limit of the sample}
\item{MDL}{The analytical method detection limit of the sample}
\item{Units}{The units of the result}
\item{LabComments}{Comments provided by the analytical laboratory}
\item{MME_Comments}{Comments provided by staff from the Mercury Monitoring and
    Evaluation section}
\item{QualCode}{Qualification code(s) for the sample. "J" indicates that there
    was an issue with the sample, but it is still being used in analyses. "R"
    indicates that there was an issue with the sample, and it is not being used
    in analyses. "BD" indicates that there was a detection in a blank sample
    associated with the sample. "FV" indicates that there was a RPD value greater
    than its acceptable limit in a field duplicate pair associated with the
    sample, which could be due to variablity in the sampling method or the
    water being sampled.}
}}
\usage{
conc_data
}
\description{
Contains all water concentration data collected for the Yolo Bypass
    Mass Balance Study. The raw data for this dataset were integrated,
    cleaned, and QA'ed in the "Compile_and_Clean_Conc_Data.R" script file.
}
\keyword{datasets}
