% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loads_calc}
\alias{loads_calc}
\title{Calculated Mass Loads}
\format{
data frame with 1,460 rows and 10 columns
\describe{
\item{SamplingEvent}{The sampling event for the mass load value}
\item{Year}{The year of the flood}
\item{StationName}{The station name}
\item{Analyte}{The analytical parameter for the mass load value}
\item{Conc}{The water concentration of the analyte used in the load calculation}
\item{ConcUnits}{The units of the \code{Conc} variable}
\item{Flow}{The daily average flow value in cubic feet per second used in the
    load calculations}
\item{LocType}{A categorical variable indicating whether the station was an
    inlet, outlet, or Below Liberty Island location}
\item{Load}{The calculated mass load value}
\item{LoadUnits}{The units of the load value}
}
}
\usage{
loads_calc
}
\description{
Contains calculated mass loads for the Yolo Bypass Mass Balance
    Study. The calculations for this dataset are in the
    "YB_Mass_Balance/Loads/All_LoadCalcs.R" script file.
}
\keyword{datasets}
