% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loads_flow_cf}
\alias{loads_flow_cf}
\title{Flow and Mass Load Data from the CALFED 2008 report}
\format{
data frame with 14 rows and 7 columns
\describe{
\item{SampleDate}{The date of the total inflow and net load values}
\item{Year}{The year of the flood}
\item{Analyte}{The analytical parameter for the net load value}
\item{Segment}{The geographical segment for the net load value}
\item{LoadUnits}{The units of the net load value}
\item{NetLoad}{The net load value}
\item{TotalInputFlow}{The total daily inflow value in cubic feet per second}
}
}
\usage{
loads_flow_cf
}
\description{
Contains total daily inflow and net methylmercury load data from
    the 2006 flood event in the Yolo Bypass as presented in the CALFED 2008 report.
    This dataset is used as a comparison to the net MeHg loads from the Yolo Bypass
    Mass Balance study.
}
\keyword{datasets}
